feat: Support AA Large and Fail (#57)
BREAKING CHANGE: alters the output of failing tests. If you're using the empty string as a failure status, please adjust to looking for 'Fail' instead.

This adds `AA Large` from 3-4.5 and returns `'Fail'` when under 3 now. The empty string I think could be misconstrued as an error, since the output is not explicit.